pos: tell the teller the token died, instead of looking offline
Two bugs with one shape: the till could know its bearer had been refused
and never say so.

The re-login banner was gated on the cached JWT having passed its `exp`.
But expiry is only one of the ways a token dies — a rotated signing
secret, a deactivated user, a suspended org all get refused while `exp`
is still hours away. In that case the outbox parked, `sync_status`
reported nothing, and the till looked merely offline. Worse, the
recovery path un-parked on the same reasoning: unexpired, therefore the
401 must have been a portal blip. So it un-parked, drained, took another
401, re-parked, and went round again until the token finally expired.
That wait is what the teller experienced.

The flag needed no gate. `auth_paused` is latched from exactly one place
— a `CoreError::Unauthenticated`, which `status_to_error` produces only
for a 401 carrying our own error envelope; a captive portal answers with
HTML and is classified `Offline` well before that. It already means "our
backend read this bearer and refused it", which is stronger than any
inference from the clock. So the banner trusts it (still only while
online, since re-auth needs a server to mint from), and un-parking now
ASKS: one authenticated request, un-park only if the backend accepts.

And the host raised the prompt on the connectivity edge, so a token
refused mid-shift — with the till online throughout — never crossed one.
It raises on the PARK edge now, which covers the restore case too: the
core reports `authPaused` as false while offline, so coming back online
flips it false → true there.

CI: `flutter analyze` failed on a no-op `.toInt()` on a `PlatformInt64`,
and three checkout files were unformatted. Both fixed.

+++ packages/features/checkout/lib/src/checkout_provider.dart
@@ -205,8 +205,8 @@ class CheckoutState {
         priceMinor: 0,
         costCurrency: loyaltyMember?.mode ?? 'points',
         costAmount: loyaltyAnyItemCost,
-        costLabel:
-            '$loyaltyAnyItemCost ${loyaltyMember?.balanceLabel ?? ''}'.trim(),
+        costLabel: '$loyaltyAnyItemCost ${loyaltyMember?.balanceLabel ?? ''}'
+            .trim(),
       );
     }
     return null;
@@ -404,7 +404,9 @@ class CheckoutNotifier extends Notifier<CheckoutState> {
           loyaltyMember: scan.member,
           loyaltyRewards: scan.rewards,
           loyaltyAnyItem: scan.anyItem,
-          loyaltyAnyItemCost: scan.anyItemCost.toInt(),
+          // `PlatformInt64` is `int` on every platform this app ships to, so
+          // the conversion the analyzer flags here would be a no-op.
+          loyaltyAnyItemCost: scan.anyItemCost,
           loyaltyError: null,
         ),
       );
